requirements.txt
aiogram==3.*
python-dotenv
sentence-transformers
torch
numpy
scikit-learn
